fix: Make it so the id field actually works now
Wow. So it turns out that the issue was me taking in a temp variable
thet gets destroyed when I exit the function.

Thus, I make an empty struct and read directly into the struct. The fact
that it would read is UB, appraently.

I also updated the macro name to reflect something a little more
accurate, and the field names no longer have a chunk_ prefix

TODO: Update the rest of the code to use this. and add a
read_chunk_and_validate function